requirements.txt
arrow==1.3.0
asgiref==3.8.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
diff-match-patch==20230430
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-dotenv==1.4.2
django-environ==0.11.2
django-import-export==4.0.6
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.1
gevent==24.2.1
greenlet==3.0.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
redis==5.0.4
rq==1.16.2
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
times==0.7
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
zope.event==5.0
zope.interface==6.4.post1
